基本完成妈妈的自动下载功能

+++ src/main/java/com/example/helloworld/AutoDown2.java
@@ -1,13 +1,10 @@
 package com.example.helloworld;
 
-import cn.hutool.http.HttpRequest;
-import cn.hutool.http.HttpResponse;
-import cn.hutool.http.HttpUtil;
-import cn.hutool.json.JSON;
 import cn.hutool.json.JSONArray;
 import cn.hutool.json.JSONObject;
 import cn.hutool.json.JSONUtil;
 
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.jsoup.Jsoup;
@@ -17,12 +14,11 @@ import org.jsoup.select.Elements;
 
 import java.io.File;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.function.Predicate;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
 
 /**
  * @author 马克
@@ -30,38 +26,65 @@ import java.util.regex.Pattern;
  */
 @Slf4j
 public class AutoDown2 {
+    private static final String ROOT_DIR = "/Users/mark/Documents/妈妈下载/";
+    private static final Map<String, String> NUMBER_FILENAME_TO_URL = new HashMap<>();
+    private static final Map<String, String> CHINESE_FILENAME_TO_URL = new HashMap<>();
+
+
     public static void main(String[] args) throws Exception {
         // 请求url
         String url = "http://api.yageapp.com/api/web/share/postor.php?aid=37759&sid=1621455&bundleid=&base_uid=1480873";
         Document document = Jsoup.connect(url).get();
+        String title = document.title();
 
         // 取中文文件名
-        List<String> chineseFileName = new ArrayList<>();
+        List<String> chineseFilename = new ArrayList<>();
         Elements elements = document.getElementsByClass("qui_list__txt");
-        elements.forEach(e -> chineseFileName.add(e.text()));
+        elements.forEach(e -> chineseFilename.add(e.text()));
 
+        // 获取数字文件名
         List<String> numName = getNumName(document);
-        Map<String, String> nameToUrl = getNameToUrl(document, numName);
+        // 获取 fileNameToUrl
+        getNameToUrl(document, numName);
+        // 处理下载信息 chineseFilename to url
+        transformNumberNameToChineseName(chineseFilename, numName);
+        // 将要下载的文件
+        List<File> targetFiles = shallCreateFile(title);
+        // 下载文件
+        targetFiles.forEach( e -> downloadFile(e, CHINESE_FILENAME_TO_URL));
 
-        Map<String, String> downloadInfo = transformName(chineseFileName, numName, nameToUrl);
-        for (String s : downloadInfo.keySet()) {
-            File file = new File(s);
-            URL fileUrl = new URL(downloadInfo.get(s));
-            FileUtils.copyURLToFile(fileUrl, file);
-            Thread.sleep(2000L);
-            log.info("{} download done.", s);
-        }
     }
 
-    public static Map<String, String> transformName (List<String> chineseFileName, List<String> numName, Map<String, String> numNameToUrl) {
-        Map<String, String> result = new HashMap<>();
-        for (int i = 0; i < numName.size(); i++) {
-            String url = numNameToUrl.get(numName.get(i));
-            String suffixName = url.substring(url.lastIndexOf("."));
-            String fileName = chineseFileName.get(i);
-            result.put(fileName + suffixName, numNameToUrl.get(numName.get(i)));
+
+    @SneakyThrows
+    public static void downloadFile(File file, Map<String, String> filenameToUrl) {
+        if (file.exists()) {
+            log.info("{} file already exist, skip and continue", file.getName());
+            return;
+        }
+        URL fileUrl = new URL(filenameToUrl.get(file.getName()));
+        FileUtils.copyURLToFile(fileUrl, file);
+        Thread.sleep(2000L);
+        log.info("{} download done.", file.getName());
+    }
+
+    public static List<File> shallCreateFile(String title) {
+        Set<String> strings = CHINESE_FILENAME_TO_URL.keySet();
+        List<File> result = new ArrayList<>();
+        for (String fileName : strings) {
+            File file = new File(ROOT_DIR.concat(title).concat(File.separator).concat(fileName));
+            result.add(file);
         }
         return result;
+    }
+
+    public static void transformNumberNameToChineseName(List<String> chineseFileName, List<String> numName) {
+        for (int i = 0; i < numName.size(); i++) {
+            String url = NUMBER_FILENAME_TO_URL.get(numName.get(i));
+            String suffixName = url.substring(url.lastIndexOf("."));
+            String fileName = chineseFileName.get(i);
+            CHINESE_FILENAME_TO_URL.put(fileName + suffixName, NUMBER_FILENAME_TO_URL.get(numName.get(i)));
+        }
     }
 
     public static List<String> getNumName (Document document) {
@@ -73,15 +96,13 @@ public class AutoDown2 {
         return midListJson.toList(String.class);
     }
 
-    public static Map<String, String> getNameToUrl(Document document, List<String> numNameList) {
+    public static void getNameToUrl(Document document, List<String> numNameList) {
         // 读取js
         Element script = document.select("script").get(7);
         // 获取文件url
         String pathArrayStr = extractVariableValue(script.html(), "patharray");
         JSONObject jsonObject = JSONUtil.parseObj(pathArrayStr);
-        Map<String, String> result = new HashMap<>();
-        numNameList.forEach(e -> result.put(e, jsonObject.getJSONObject(e).getStr("playurl")));
-        return result;
+        numNameList.forEach(e -> NUMBER_FILENAME_TO_URL.put(e, jsonObject.getJSONObject(e).getStr("playurl")));
     }
 
     public static String extractVariableValue(String content, String variableName) {
@@ -92,4 +113,6 @@ public class AutoDown2 {
         }
         return null;
     }
+
+    static <T> Predicate<T> not(Predicate<T> p) { return o -> !p.test(o); }
 }
